app/server/schemas.py
from typing import Optional, List

from pydantic import BaseModel, EmailStr, Field
from bson import ObjectId



class StudentSchema(BaseModel):
    fullname: str = Field(...)
    email: EmailStr = Field(...)
    course_of_study: str = Field(...)
    year: int = Field(..., gt=0, lt=18)
    gpa: float = Field(..., le=4.0)

    class Config:
        schema_extra = {
            "example": {
                "fullname": "John Doe",
                "email": "jdoe@x.edu.ng",
                "course_of_study": "Water resources engineering",
                "year": 16,
                "gpa": "3.0",
            }
        }



class StudentOutSchema(StudentSchema):
    id: int 
   
# class ManyStudentsInResponse(StudentOutSchema):
#     students: List[StudentOutSchema] = []
#     students_count: int = Schema(..., alias="studentsCount")
   

class UpdateStudentSchema(BaseModel):
    fullname: Optional[str]
    email: Optional[EmailStr]
    course_of_study: Optional[str]
    year: Optional[int]
    gpa: Optional[float]

    class Config:
        schema_extra = {
            "example": {
                "fullname": "John Doe",
                "email": "jdoe@x.edu.ng",
                "course_of_study": "Water resources and environmental engineering",
                "year": 16,
                "gpa": "4.0",
            }
        }


def ResponseModel(data, message):
    return {
        "data": [data],
        "code": 200,
        "message": message,
    }


def ErrorResponseModel(error, code, message):
    return {"error": error, "code": code, "message": message}



