requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
pydantic[email]
motor==2.5.1
python-decouple==3.5